[style]pin share 코드 분리
