+++ .eslintrc
@@ -13,7 +13,8 @@
         "endOfLine": "auto"
       }
     ],
-    "react/react-in-jsx-scope": "off"
+    "react/react-in-jsx-scope": "off",
+    "react/jsx-uses-react":"off"
   },
   "settings": {
     "react": {
